[Add] logic to install.sh

+++ install.sh
@@ -1,12 +1,11 @@
 # /bin/sh
 
-ln -sf $HOME/.settings/zsh/.zshenv $HOME/.zshenv
-ln -sf $HOME/.settings/vim/vimrc $HOME/.vimrc
+[ -e "$HOME/.zshenv" ]&& ln -sf $HOME/.settings/zsh/.zshenv $HOME/.zshenv
+[ -e "$HOME/.vimrc" ]&& ln -sf $HOME/.settings/vim/vimrc $HOME/.vimrc
 
 [ "$SHELL" != "$(which zsh)" ]&& chsh -s $(which zsh)
 cd $HOME/.settings 
-    git submodule init
-    git submodule update
+    git submodule update --recursive --init
 cd $HOME
 
 [ ! -f "$HOME/.zshlocal" ]&& echo "#External zsh local settings" >"$HOME/.zshlocal"
